+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Updated webview login with version handshake](https://github.com/multiversx/mx-sdk-dapp/pull/1460)
+
 ## [[5.0.0-alpha.8](https://github.com/multiversx/mx-sdk-dapp/pull/1459)] - 2025-06-19
 
 - [Added `getScamFlag` utility](https://github.com/multiversx/mx-sdk-dapp/pull/1458)